BLE config ready

+++ Projects/DataProviderApp/app/src/main/java/com/example/dataproviderapp/api/OBDIIPIDSApi.kt
@@ -7,5 +7,5 @@ import retrofit2.http.Path
 
 interface OBDIIPIDSApi {
     @GET("obdii-pids/current-data-service/{supportedPidsUint}")
-    suspend fun getCurrentDataSupportedPids(@Path("supportedPidsUint") pids: UInt): Response<PidsDetailDto>
+    suspend fun getCurrentDataSupportedPids(@Path("supportedPidsUint") pids: Long): Response<PidsDetailDto>
 }
