+++ Projects/DataProviderApp/app/src/main/java/com/example/dataproviderapp/dto/responses/PidsDetailDto.kt
@@ -1,6 +1,6 @@
 package com.example.dataproviderapp.dto.responses
 
 data class PidsDetailDto(
-    val once: List<UShort>,
-    val repeatable: List<UShort>
+    val once: List<Int>,
+    val repeatable: List<Int>
 )
