+++ Projects/DataProviderApp/app/src/main/java/com/example/dataproviderapp/ui/Nav/NavViewModel.kt
@@ -617,7 +617,7 @@ class NavViewModel : ViewModel() {
         }
     }
 
-    fun getCurrentDataSupportedPids(pids: UInt) {
+    fun getCurrentDataSupportedPids(pids: Long) {
         viewModelScope.launch {
             _currentDataSupportedPidsState.value = CurrentDataSupportedPidsDetailsState.Loading
